move relations related files
